backend/requirements.txt
django==3.2.*
djangorestframework==3.14.*
markdown
django-filter
django-environ
django-guardian
djangorestframework-guardian
django-cors-headers
django-rest-authemail
python-dotenv
boto3
django-storages
openai==0.8.0
Pillow
torch
torchvision